fix: use 127.0.0.1 instead of 0.0.0.0 for ollama API
"0.0.0.0" is not intended for direct access by users as it
represents "any available network interface" and is often
used by applications to bind to all network addresses.
Application (ollama) is not running at "0.0.0.0".

To representing localhost, "127.0.0.1" or "localhost"
should be used. Here we use "127.0.0.1" since it's the
most reliable way to access localhost.