KirigamiSettings: CategorizedSettings -> ConfigurationView
This change fixes the General Settings pane not being shown at all.

ConfigurationView is the most recently supported way to handle the
configuration, and most if not all other applications by kde.org
switched to it. It seems it works better in all scenarios (desktop
and mobile).

The change basically required moving the building of the QML bits
to a proper module, which is why ecm_add_qml_module is now used.
Due to this change, main.qml was renamed to the uppercased Main.qml,
if I understand it correctly, to allow for loading it through
engine.loadFromModule.