ENHANCEMENT Multiline prompt
A one line prompt prevented the user from being able to see the pasted
or inserted multiline structured text correctly and prevented editing it
afterward.

This commit replaces the old single line prompt with a multiline which
initialy has a small height and then starts to grow to a cap when new
lines are added. After hitting the cap a scrollbar apears.

The prompt's newlines are preserved in the prompt balloon.